requirements.txt
numpy==1.24.3
scikit-learn==1.3.0
pillow==9.5.0
tensorflow==2.13.0
streamlit==1.24.1
scipy==1.11.1
keras==2.13.1